Move checks for docker and conda envs into check namespace of CLI
