+++ calkit/cli/check.py
@@ -2,11 +2,16 @@
 
 from __future__ import annotations
 
+import functools
+import hashlib
+import json
+import os
 import subprocess
 from typing import Annotated
 
 import typer
 
+import calkit
 from calkit.check import check_reproducibility
 from calkit.cli import raise_error
 
@@ -46,3 +51,120 @@ def check_call(
             typer.echo("Fallback call succeeded")
         except subprocess.CalledProcessError:
             raise_error("Fallback call failed")
+
+
+@check_app.command(
+    name="docker-env",
+    help="Check that Docker image is up-to-date.",
+)
+def check_docker_env(
+    tag: Annotated[str, typer.Argument(help="Image tag.")],
+    fpath: Annotated[
+        str, typer.Option("-i", "--input", help="Path to input Dockerfile.")
+    ] = "Dockerfile",
+    platform: Annotated[
+        str, typer.Option("--platform", help="Which platform(s) to build for.")
+    ] = None,
+    quiet: Annotated[
+        bool, typer.Option("--quiet", "-q", help="Be quiet.")
+    ] = False,
+):
+    def get_docker_inspect():
+        out = json.loads(
+            subprocess.check_output(["docker", "inspect", tag]).decode()
+        )
+        # Remove some keys that can change without the important aspects of
+        # the image changing
+        _ = out[0].pop("Id")
+        _ = out[0].pop("RepoDigests")
+        _ = out[0].pop("Metadata")
+        _ = out[0].pop("DockerVersion")
+        return out
+
+    outfile = open(os.devnull, "w") if quiet else None
+    typer.echo(f"Checking for existing image with tag {tag}", file=outfile)
+    # First call Docker inspect
+    try:
+        inspect = get_docker_inspect()
+    except subprocess.CalledProcessError:
+        typer.echo(f"No image with tag {tag} found locally", file=outfile)
+        inspect = []
+    typer.echo(f"Reading Dockerfile from {fpath}", file=outfile)
+    with open(fpath) as f:
+        dockerfile = f.read()
+    dockerfile_md5 = hashlib.md5(dockerfile.encode()).hexdigest()
+    lock_fpath = fpath + "-lock.json"
+    rebuild = True
+    if os.path.isfile(lock_fpath):
+        typer.echo(f"Reading lock file: {lock_fpath}", file=outfile)
+        with open(lock_fpath) as f:
+            lock = json.load(f)
+    else:
+        typer.echo(f"Lock file ({lock_fpath}) does not exist", file=outfile)
+        lock = None
+    if inspect and lock:
+        typer.echo(
+            "Checking image and Dockerfile against lock file", file=outfile
+        )
+        rebuild = inspect[0]["RootFS"]["Layers"] != lock[0]["RootFS"][
+            "Layers"
+        ] or dockerfile_md5 != lock[0].get("DockerfileMD5")
+    if rebuild:
+        wdir, fname = os.path.split(fpath)
+        if not wdir:
+            wdir = None
+        cmd = ["docker", "build", "-t", tag, "-f", fname]
+        if platform is not None:
+            cmd += ["--platform", platform]
+        cmd.append(".")
+        subprocess.check_call(cmd, cwd=wdir)
+    # Write the lock file
+    inspect = get_docker_inspect()
+    inspect[0]["DockerfileMD5"] = dockerfile_md5
+    with open(lock_fpath, "w") as f:
+        json.dump(inspect, f, indent=4)
+
+
+@check_app.command(
+    name="conda-env",
+    help="Check a conda environment and rebuild if necessary.",
+)
+def check_conda_env(
+    env_fpath: Annotated[
+        str,
+        typer.Option(
+            "--file", "-f", help="Path to conda environment YAML file."
+        ),
+    ] = "environment.yml",
+    output_fpath: Annotated[
+        str,
+        typer.Option(
+            "--output",
+            "-o",
+            help=(
+                "Path to which existing environment should be exported. "
+                "If not specified, will have the same filename with '-lock' "
+                "appended to it, keeping the same extension."
+            ),
+        ),
+    ] = None,
+    relaxed: Annotated[
+        bool,
+        typer.Option(
+            "--relaxed", help="Treat conda and pip dependencies as equivalent."
+        ),
+    ] = False,
+    quiet: Annotated[
+        bool, typer.Option("--quiet", "-q", help="Be quiet.")
+    ] = False,
+):
+    if quiet:
+        log_func = functools.partial(typer.echo, file=open(os.devnull, "w"))
+    else:
+        log_func = typer.echo
+    calkit.conda.check_env(
+        env_fpath=env_fpath,
+        output_fpath=output_fpath,
+        log_func=log_func,
+        relaxed=relaxed,
+    )
